requirements.txt
absl-py @ file:///C:/ci/absl-py_1600290345082/work
aiohttp @ file:///C:/ci/aiohttp_1602530358780/work
anyio @ file:///C:/ci/anyio_1620153167783/work/dist
appdirs==1.4.4
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
audioread==2.1.9
AutoReg==1.1
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
blinker==1.4
blis==0.7.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
catalogue==2.0.4
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1600699250966/work
chardet==3.0.4
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
comtypes==1.1.10
cryptography @ file:///C:/ci/cryptography_1601046913206/work
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
descartes==1.1.0
docopt==0.6.2
entrypoints==0.3
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.1
gast==0.3.3
google-auth @ file:///tmp/build/80754af9/google-auth_1601995530934/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GPUtil==1.4.0
grpcio @ file:///C:/ci/grpcio_1597406403308/work
gTTS==2.2.3
gunicorn==20.1.0
h5py==2.10.0
huggingface-hub==0.0.12
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617121109687/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.17.0
jieba3k==0.35.1
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
joblib==1.0.1
Js2Py==0.71
json5==0.9.5
jsonify==0.5
jsonschema==3.0.2
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213516947/work
jupyter-packaging @ file:///tmp/build/80754af9/jupyter-packaging_1613502826984/work
jupyter-server @ file:///C:/ci/jupyter_server_1616084265530/work
jupyterlab @ file:///tmp/build/80754af9/jupyterlab_1619133235951/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1617134334258/work
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
keras-datasets==0.1.0
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
lesscpy==0.15.0
librosa==0.8.1
llvmlite==0.36.0
lxml==4.6.3
Markdown @ file:///C:/ci/markdown_1603216569530/work
MarkupSafe @ file:///C:/ci/markupsafe_1621528383308/work
matplotlib==3.4.2
mistune @ file:///C:/ci/mistune_1594373272338/work
mizani==0.7.3
mkl-fft==1.2.0
mkl-random==1.1.0
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1600456486794/work
murmurhash==1.0.5
nbclassic @ file:///tmp/build/80754af9/nbclassic_1616085367084/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
newspaper3k==0.2.8
nltk==3.6.2
notebook @ file:///C:/ci/notebook_1621528807363/work
numba==0.53.1
numpy @ file:///C:/ci/numpy_and_numpy_base_1596233945180/work
oauthlib==3.1.0
opencv-python==4.5.2.54
opt-einsum==3.1.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
palettable==3.3.0
pandas @ file:///C:/ci/pandas_1618264472954/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pathy==0.6.0
patsy==0.5.1
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.2.0
pipwin==0.5.1
playsound==1.3.0
plotnine==0.8.0
ply==3.11
pooch==1.4.0
preshed==3.0.5
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1618088486455/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/jaisri/pipwin/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.7.4
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyjsparser==2.7.1
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pypiwin32==223
PyPrind==2.11.3
pyreadline==2.1
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
pySmartDL==1.3.4
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pyttsx3==2.90
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
regex==2021.4.4
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-file==1.5.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
sacremoses==0.0.45
scikit-learn==0.24.2
scipy @ file:///C:/ci/scipy_1597686737426/work
seaborn==0.11.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smart-open==5.1.0
sniffio @ file:///C:/ci/sniffio_1614030522573/work
SoundFile==0.10.3.post1
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.6
SpeechRecognition==3.8.1
srsly==2.4.1
statsmodels==0.12.2
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_baze37/tensorboard_1622025863084/work/tensorboard-2.5.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.9.4
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
thinc==8.0.6
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.10.3
torch==1.10.0.dev20210726+cpu
torchaudio==0.10.0.dev20210727
torchvision==0.11.0.dev20210727+cpu
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm==4.61.1
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
transformers==4.8.2
typer==0.3.2
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
tzdata==2021.1
tzlocal==3.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wasabi==0.8.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1602679793616/work
zipp @ file:///tmp/build/80754af9/zipp_1603306325174/work
